Fixed EIGRP Neighbor status not being a float64 and therefore no metric created

+++ metrics/eigrp_adjacency.go
@@ -69,7 +69,7 @@ func parseEigrpAdjMsg(msg *telemetry.Telemetry, dm *DeviceGroupedMetrics, t time
 
 		// Instrument EIGRP Adjacency Status
 		CreatePromMetric(
-			1,
+			float64(1),
 			eigrpAdjStatus,
 			prometheus.GaugeValue,
 			dm, t,
